chore: move commands to cli
