+++ engine/cli/commands/chat_cmd.cc
