+++ engine/cli/commands/chat_cmd.h
