+++ engine/cli/commands/chat_completion_cmd.h
