+++ engine/cli/commands/model_del_cmd.h
