+++ engine/cli/commands/model_status_cmd.h
