+++ engine/cli/commands/model_stop_cmd.h
