+++ engine/cli/commands/server_stop_cmd.cc
